# server/docs/openapi.yaml
openapi: 3.0.0
info:
  description: REST API for the NDHU campus second-hand textbook marketplace.
  title: NDHU Second-Hand Book Store API
  termsOfService: http://swagger.io/terms/
  contact:
    name: SHBS Dev Team
  license:
    name: MIT
  version: "1.0"
paths:
...
  /listings:
    get:
      description: Paginated, filterable list of book listings. Returns the full page
        in the JSON body and total count in X-Total-Count header.
      tags:
        - Listings
      summary: List listings
      operationId: getListings
      parameters:
        - description: "Filter by status (default: active)"
          name: status
          in: query
          schema:
            type: string
        - description: Filter by department
          name: department
          in: query
          schema:
            type: string
        - description: Filter by condition (good|moderate|poor)
          name: condition
          in: query
          schema:
            type: string
        - description: Full-text search on title, author, or ISBN
          name: search
          in: query
          schema:
            type: string
        - description: Minimum price
          name: price_min
          in: query
          schema:
            type: number
        - description: Maximum price
          name: price_max
          in: query
          schema:
            type: number
        - description: "Page number (default: 1)"
          name: page
          in: query
          schema:
            type: integer
        - description: "Items per page (default: 20, max: 100)"
          name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/model.ListingWithImages"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
    post:
      security:
        - BearerAuth: []
      description: Publishes a new book listing for the authenticated seller.
      tags:
        - Listings
      summary: Create listing
      operationId: createListing
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/model.SwaggerCreateListingRequest"
        description: Listing details
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.ListingWithImages"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
  /listings/me:
    get:
      security:
        - BearerAuth: []
      description: Returns all listings created by the authenticated user, newest first.
      tags:
        - Listings
      summary: My listings
      operationId: getMyListings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/model.ListingWithImages"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
  "/listings/{id}":
    get:
      description: Returns a single listing with seller info and image URLs.
      tags:
        - Listings
      summary: Get listing by ID
      operationId: getListing
      parameters:
        - description: Listing ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.ListingWithImages"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
    delete:
      security:
        - BearerAuth: []
      description: Soft-deletes a listing by setting its status to 'delisted'. Only
        the seller or an admin may delete.
      tags:
        - Listings
      summary: Delete listing
      operationId: deleteListing
      parameters:
        - description: Listing ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
    patch:
      security:
        - BearerAuth: []
      description: Partially updates a listing. Only the seller may update their own
        listing.
      tags:
        - Listings
      summary: Update listing
      operationId: updateListing
      parameters:
        - description: Listing ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/model.SwaggerUpdateListingRequest"
        description: Fields to update
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.ListingWithImages"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
...
servers:
  - url: /api/v1
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    model.ListingWithImages:
      type: object
      properties:
        ai_confidence:
          type: number
        author:
          type: string
        condition:
          type: string
        course_code:
          type: string
        created_at:
          type: string
        department:
          type: string
        description:
          type: string
        id:
          type: string
        image_urls:
          type: array
          items:
            type: string
        isbn:
          type: string
        price:
          type: number
        seller_avatar:
          type: string
        seller_id:
          type: string
        seller_name:
          type: string
        status:
          type: string
        title:
          type: string
        updated_at:
          type: string
...
    model.SwaggerCreateListingRequest:
      type: object
      properties:
        author:
          type: string
        condition:
          type: string
        course_code:
          type: string
        department:
          type: string
        description:
          type: string
        image_ids:
          type: array
          items:
            type: string
        isbn:
          type: string
        price:
          type: number
        title:
          type: string
    model.SwaggerErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    model.SwaggerUpdateListingRequest:
      type: object
      properties:
        author:
          type: string
        condition:
          type: string
        course_code:
          type: string
        department:
          type: string
        description:
          type: string
        image_ids:
          type: array
          items:
            type: string
        isbn:
          type: string
        price:
          type: number
        status:
          type: string
        title:
          type: string
...